backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0  # PostgreSQL
aiosqlite>=0.19.0  # SQLite

# LLMs
anthropic>=0.18.0
openai>=1.12.0
langchain>=0.1.6
langgraph>=0.0.20
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0

# Tools
tavily-python>=0.3.0
beautifulsoup4>=4.12.3
httpx>=0.26.0
lxml>=5.1.0

# Code execution
RestrictedPython>=6.2
matplotlib>=3.8.2
plotly>=5.18.0
pandas>=2.2.0
numpy>=1.26.3

# SSE
sse-starlette>=1.8.2

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
